requirements.txt
streamlit
torch
torchvision
opencv-python
Pillow
git+https://github.com/openai/CLIP.git
ultralytics
numpy==1.24.4
